Move `GrpcTestUtil` into same package as `GrpcUtil`
